pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "microsoft-365-emad"
version = "0.1.0"
description = "Microsoft 365 Service Broker eMAD — email, calendar, OneDrive via Graph API"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.60",
    "langchain-core>=0.3.28",
    "langchain-openai>=0.2.14",
    "msal>=1.31.1",
    "httpx>=0.27.2",
    "pydantic>=2.10.0",
    "asyncpg>=0.30.0",
    "pyyaml>=6.0.2",
    "prometheus-client>=0.21.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
microsoft_365_emad = ["prompts/*.md"]
